+++ tool_packages/unique_web_search/src/unique_web_search/services/search_engine/utils/grounding/vertexai/__init__.py
@@ -1,12 +1,9 @@
 from unique_web_search.services.search_engine.utils.grounding.vertexai.client import (
     get_vertex_client,
 )
-from unique_web_search.services.search_engine.utils.grounding.vertexai.config import (
-    get_vertex_grounding_config,
-    get_vertex_grounding_with_structured_output_config,
-)
 from unique_web_search.services.search_engine.utils.grounding.vertexai.gemini import (
     generate_vertexai_response,
+    get_vertex_grounding_with_structured_output_config,
 )
 from unique_web_search.services.search_engine.utils.grounding.vertexai.response_handler import (
     add_citations,
@@ -14,7 +11,6 @@ from unique_web_search.services.search_engine.utils.grounding.vertexai.response_
 
 __all__ = [
     "get_vertex_client",
-    "get_vertex_grounding_config",
     "get_vertex_grounding_with_structured_output_config",
     "generate_vertexai_response",
     "add_citations",
